Update starrers list
Thanks for the stars! I think that soon, however, I will stop posting
starrers as otherwise the list will be too long, and I will have to
update it continously (not that that would be a bad thing :)) so
apologies if I don't update the starrers list as frequently from now on.